Add Literal Data Meta Hash subpacket.
Rationale: This helps for the rare case that the meta data (i.e. the
file name) gives important information (think "code-for-go.txt",
"code-for-scrub.txt") for an action.  Sure the authenticated
encryption (CFB.MDC or OCB) protects against this but it might be
worth to also protect the case of a leaked signed message which get
re-encrypted.