Rename the Additional Encryption Subkey key flag
Rationale: It is not useful to base the ADSK feature just on a key
flag because this is not flexible enough and does not handle well the
case of subkey-rollover or multiple.  This is better does with further
meta data. However the general idea of having an encryption capable
subkey which is used only in a restricted setting is indeed useful.
The new name might not be the best one but it allows to use "r" as
short identifier in key listings.