v6 trailer uses 4 octets, like v4 trailer
The only variable-length field in the trailer is the hashed
subpackets, which itself permits only a 32-bit length.  The
fixed-length fields in the trailer are fairly small, so it's very
unlikely that this trailer will overflow.  8 octets is overkill, and
keeping them makes it possible to have the v6 hashed data stream alias
to the stream that is signed with a v3 signature.

Closes: #130