Fix composition of public key blocks.
In the course of the reformatting actions of the draft a regression
against 4880 was not fixed (Zero User ID packets).  The reason for
introducing zero User ID packets might have been the idea to express
that an Attribute packet may be used instead of a User ID.  However,
that should either be clarified in the comments or left to the
implementation.

The second fix is to require at least one Signature packet after a
User ID and Attribute packet.  This was wrong in 2440 and 4880 but is
cryptographically required.