Update "Confidentiality via Encryption" section
The first paragraph and the sequence that follows it refer to using
public-key encryption. This updates the first two lines to reflect
that.

Then, the first mentioned method for symmetric encryption
(encrypting the message with a key derived from a shared secret
directly, rather than encrypting a session key with a derived key)
is deprecated, so we don't mention it here.

Finally, simplify and clean up the language slightly.