Forbid primary keys that cannot sign.
This addresses an open question from Jonathan Hammell in
https://mailarchive.ietf.org/arch/msg/openpgp/pr7OhX5Qh9yzQ0-ZZxxDC3HgzcY
by taking an opinionated stance that encryption-only primary keys are not
acceptable and will be rejected.

The other possible resolution is to say that an encryption-only primary key
is fine because other implementations MAY sign relevant things like direct key
signatures.

However, that would mean that the primary key can't indicate algorithm
preferences, expiration dates, or bind subkeys in a way that any other
implementation can make sense of without a trust model about who gets to
select these preferences.

the result of that other choice seems to be an encryption-only OpenPGP
certificate with indefinite (infinite?) lifespan that can only use the
MTI algorithms.  I doubt this is a useful thing to have in the ecosystem.