Confidentiality via Encryption: indicate that the session key may not be used directly
This update tries to incorporate the insight Falko Strenzke originally offered in !265.

Here, we try to apply it on top of Daniel Huigens' "Editorial Changes" MR !261.

For simplicity of language, i've cleaned up "the session key or,
depending on packet version, a message key derived from the session
key" and replaced it with just "a message key derived from the session
key".

for v3 PKESK/v1 SEIPD this is still trivially correct, as the identity
function is used for the derivation ☺ -- but this way the text is
simpler and clearer, and it focuses more on the modern variant.