Don't encrypt symmetric key algorithm ID in X25519 and X448
Leave the symmetric key algorithm identifier in plaintext, and prepend
it to the ciphertext of the session key, so as to give AES Key Wrap an
input that is a multiple of 8 octets.

Since the session key size can't be obfuscated in the X25519 and X448
algorithms, encrypting the symmetric algorithm identifier offers little
added value.