Move the handling of the symmetric algorithm ID into the public-key algorithms
Rather than concatenating the symmetric algorithm identifier and the
session key in the v3 PKESK packet description, and then splitting it
up again in the X25519 and X448 algorithms, pass the symmetric algorithm
and session key separately, and let the public-key algorithms define how
to concatenate or otherwise use them.

Additionally, fix the description of the session key checksum to not
include the symmetric algorithm identifier.