X25519 and X448: mix public key material into KDF
RFC 7748 recommends ensuring that the key-derivation function includes
both publc keys as well as the shared secret, which the current
specification does not do.  With this change, the approach for using
the CFRG curves for ECDH (X25519 and X448) now more closely follows
the recommendations in RFC 7748.

In particular, RFC 7748 says:

   Alice and Bob can then use a key-derivation function that includes
   K, K_A, and K_B to derive a symmetric key.

and:

   Designers using these curves should be aware that for each public
   key, there are several publicly computable public keys that are
   equivalent to it, i.e., they produce the same shared secrets.  Thus
   using a public key as an identifier and knowledge of a shared secret
   as proof of ownership (without including the public keys in the key
   derivation) might lead to subtle vulnerabilities.

I don't know how to produce a specific OpenPGP vulnerability from this
situation, but it can also be challenging to predict how and where
OpenPGP will be used in the future.  Following the CFRG guidance
should reduce the risk of a surprising failure.