Put Signature types in a registry table
https://www.iana.org/assignments/pgp-parameters/pgp-parameters.xhtml
currently has no registry of signature types.

However, the list of signature types is very much relevant to how
OpenPGP implementations can function.  See for example
https://gitlab.com/dkg/openpgp-stateless-cli/-/issues/69 and
https://gitlab.com/openpgp-wg/rfc4880bis/-/merge_requests/60

The WG should be explicitly asking IANA to record this as a
registry. Putting it in a table labeled "registry" is the simplest way
to do that.

This commit makes no substantive changes to the document's text, but
it does create a new table intended to be a registry tracked by IANA.