Clean up bitfield/flag registries
We have three registries of variable-length bitfields (flags) in the spec.

In order to help IANA document the registries in the most sensible way, we should
have them written in a standardized form.

This change does nothing to the semantics or contents of the tables, it just
presents them in a simple, predictable way.