Adjust v6 PKESK (X25519) material, describing both encryption and decryption
The other SEIPD sections contain enough information to both generate and consume the
examples because they rely on passwords.

To allow both generation and consumption of the PKESK sample, we document
the ephemeral secret key used in the calculation, even though it doesn't
appear on the wire.

We also explain more explicitly what the parts from the recipient key are.