Forbid generating packets with weak hash functions in S2K KDFs.
There are two motivations for this.  First, we want to phase out these
algorithms as fast as possible, and avoiding their use anywhere is a
necessary first step.  Second, we forbid the use in the ECDH KDF, so
for consistency we should also forbid using them in S2K KDFs.

Sadly, we cannot reasonably forbid consumption of such artifacts in
general, because up to version 2.4.1, GnuPG defaulted to SHA-1 as hash
function in S2K objects.  We can, however, forbid the consumption of
version 6 packets, because there are no legacy deployments that
created such artifacts.

Fixes #165.