Simplify v6 one-pass-signature packets.
Since only v6 keys can create v6 signatures, avoid hard-coding the key
version and fix the length of the fingerprint.  This avoids two warts,
first the redundant key version (which the document pointed out must
be six), and the fact that we don't know the length of fingerprints of
unknown key versions needlessly complicating parsing.

Also update the test vector.  The literal data packet uses the unicode
format octet, so this change obsoletes !311.