Remove the SaltedHash Armor Header.
The SaltedHash Header proved to be problematic, as earlier versions of
the spec did not allow headers besides the Hash header, making it
impossible to sign a cleartext signature framework message using both
a v4 and v6 key, which is important to provide a migration path to v6.

Instead, we remove the SaltedHash header, and make the Hash header
itself optional (for v4 signatures, MUST NOT for v6).  The algorithm
and salt must then be read from the signature, which prevents
one-pass-processing of the message, but as the CSF is meant for human
readable text only, we think this is an acceptable trade-off.