Drop sample ASCII-armored message (ZIP-compressed, "for eyes only")
The sample ASCII-armored message is a ZIP-compressed literal data
packet that claims its filename is "_CONSOLE".  But ZIP isn't a high
priority compression algorithm, and _CONSOLE as a filename (meaning
"for eyes only") is explicitly deprecated.  Even if it wasn't
deprecated, it's deeply weird to say "here is a cleartext message for
your eyes only".

It's also unclear in the OpenPGP ecosystem where such a packet
sequence would be useful.  For example, it's not encrypted, and it's
not signed.  It's not even well-compressed, since it's got
ASCII-armoring added back onto it.

SOP doesn't even have a datatype to be able to consume or generate
this packet sequence, no one has asked for it in SOP, and i can't
imagine a good argument for adding it.

If anything this particular ASCII-armored message should exist in a
bestiary of "things the OpenPGP spec allows you to create but you
really probably shouldn't".  I really don't think it belongs as the
first example of an ASCII-armored message.

By contrast, the rest of the draft contains many useful ASCII-armored
OpenPGP test vectors, so i don't think we lose anything with just
cutting this one out.