Change AEAD Secret-Key encryption mode.
It is unlikely that there is already an implementation using AEAD in
the format described earlier.  Even if such an implementation exists
it can't be in wide-spread use and thus I believe it is okay to change
the specification to also include the public key as AD to detect
attacks based on the ability to modify public key parameters.