Clean up terminology about "Certificate"
I scanned the draft for terminology related to the words
Certificate, Certify, and Certification.

In particular, this fixes places where "certificate" is used to mean
both "detached signature" and "User ID certification".

This change introduces no substantive difference (TPKs are already described
as being called "certificates" elsewhere), but it improves and clarifies the text.