Change the IPR from trust200902 to pre5378Trust200902
RFCs published before RFC5378 (Nov 2008) did not automatically
grant rights to use the text (for incorporation to new RFCs) to
the IETF Trust. RFC4880 was published before this date, so we are
required to request permission from all of the authors of RFC4880
for permission to include some of their text. Unfortunately, one
of tne of the authors, Hal Finney, has passed away.

Changing the IPR policy from trust200902 to pre5378Trust200902
adds a disclaimer to the IPR boiler plate to reflect this.