Remove a fixme regarding the fixed info
Including the packet type (either SEIPD or OCBED+its_version) will be
problematic for the decryption case:  We would need to parse all PKESK
and then the header of the SEIPD packet to get this info.  Only then
we can start to decrypt the PKESK.  With hundreds of PKESK we can't
start working on them immediately but need to wait for the SEIPD.

It is far easier to provide an option to reject broken or unwanted
SEIPD/OCPED versions.