Drop the remains of support for V6 signatures.
While trying to implementing V6 signatures with the restriction of a
zero-length salt, it turned out that V6 signatures do not hash the
meta data of the Literal Data packet.  This is a regression towards
the fix introduced with V5 and it makes no sense at all to
re-introduce this bug.  Thus no support for V6 signature in LibrePGP.

If there is ever a need for a larger are for signature subpackets a
new signature format needs to be introduced with just the required
changes.  Or we re-purpose parts of the unhashed data area for hashed
data using a new hash signature subpacket.